requirements.txt
pandas
matplotlib
python-docx
numpy
tqdm